+++ cpp/lib/src/DbConnection.cc
@@ -22,7 +22,10 @@
 #include <cstdlib>
 #include "FileUtil.h"
 #include "IniFile.h"
+#include "MySQLDbConnection.h"
+#include "PostgresDbConnection.h"
 #include "RegexMatcher.h"
+#include "Sqlite3DbConnection.h"
 #include "StringUtil.h"
 #include "UrlUtil.h"
 #include "UBTools.h"
@@ -197,35 +200,6 @@ void DbConnection::queryOrDie(const std::string &query_statement) {
     if (not query(query_statement))
         LOG_ERROR("in DbConnection::queryOrDie: \"" + query_statement + "\" failed: " + getLastErrorMessage());
 }
-
-
-namespace {
-
-
-enum ParseState { NORMAL, IN_DOUBLE_DASH_COMMENT, IN_C_STYLE_COMMENT, IN_STRING_CONSTANT };
-
-
-// A helper function for SplitSqliteStatements().
-// CREATE TRIGGER statements end with a semicolon followed by END.  As we usually treat semicolons as statement
-// separators we need special handling for this case.
-void AddStatement(const std::string &statement_candidate, std::vector<std::string> * const individual_statements) {
-    static RegexMatcher *create_trigger_matcher(
-        RegexMatcher::RegexMatcherFactoryOrDie("^CREATE\\s+(TEMP|TEMPORARY)?\\s+TRIGGER",
-                                               RegexMatcher::ENABLE_UTF8 | RegexMatcher::CASE_INSENSITIVE));
-
-    if (individual_statements->empty())
-        individual_statements->emplace_back(statement_candidate);
-    else if (::strcasecmp(statement_candidate.c_str(), "END") == 0 or ::strcasecmp(statement_candidate.c_str(), "END;") == 0) {
-        if (individual_statements->empty() or not create_trigger_matcher->matched(individual_statements->back()))
-            individual_statements->emplace_back(statement_candidate);
-        else
-            individual_statements->back() += statement_candidate;
-    } else
-        individual_statements->emplace_back(statement_candidate);
-}
-
-
-} // unnamed namespace
 
 
 void DbConnection::queryFileOrDie(const std::string &filename) {
@@ -739,564 +713,6 @@ std::unordered_set<DbConnection::MYSQL_PRIVILEGE> DbConnection::mySQLGetUserPriv
     }
 
     return {};
-}
-
-
-MySQLDbConnection::MySQLDbConnection(const TimeZone time_zone)
-    : MySQLDbConnection(IniFile(DEFAULT_CONFIG_FILE_PATH), "Database", time_zone)
-{
-}
-
-
-MySQLDbConnection::MySQLDbConnection(const std::string &mysql_url, const Charset charset, const TimeZone time_zone) {
-    static RegexMatcher * const mysql_url_matcher(
-        RegexMatcher::RegexMatcherFactory("mysql://([^:]+):([^@]+)@([^:/]+)(\\d+:)?/(.+)"));
-    std::string err_msg;
-    if (not mysql_url_matcher->matched(mysql_url, &err_msg))
-        LOG_ERROR("\"" + mysql_url + "\" does not look like an expected MySQL URL! (" + err_msg + ")");
-
-    const std::string user(UrlUtil::UrlDecode((*mysql_url_matcher)[1]));
-    const std::string passwd((*mysql_url_matcher)[2]);
-    const std::string host(UrlUtil::UrlDecode((*mysql_url_matcher)[3]));
-    const std::string db_name(UrlUtil::UrlDecode((*mysql_url_matcher)[5]));
-
-    const std::string port_plus_colon((*mysql_url_matcher)[4]);
-    unsigned port;
-    if (port_plus_colon.empty())
-        port = MYSQL_PORT;
-    else
-        port = StringUtil::ToUnsigned(port_plus_colon.substr(0, port_plus_colon.length() - 1));
-
-    init(db_name, user, passwd, host, port, charset, time_zone);
-}
-
-
-MySQLDbConnection::MySQLDbConnection(const IniFile &ini_file, const std::string &ini_file_section, const TimeZone time_zone) {
-    const auto db_section(ini_file.getSection(ini_file_section));
-    if (db_section == ini_file.end())
-        LOG_ERROR("DbConnection section \"" + ini_file_section +"\" not found in config file \"" + ini_file.getFilename()
-                  + "\"!");
-
-    const std::string host(db_section->getString("sql_host", "localhost"));
-    const std::string database(db_section->getString("sql_database"));
-    const std::string user(db_section->getString("sql_username"));
-    const std::string password(db_section->getString("sql_password"));
-    const unsigned port(db_section->getUnsigned("sql_port", MYSQL_PORT));
-
-    const std::map<std::string, int> string_to_value_map{
-        { "UTF8MB3", UTF8MB3 },
-        { "UTF8MB4", UTF8MB4 },
-    };
-    const Charset charset(static_cast<Charset>(db_section->getEnum("sql_charset", string_to_value_map, UTF8MB4)));
-
-    init(database, user, password, host, port, charset, time_zone);
-}
-
-
-MySQLDbConnection::~MySQLDbConnection() {
-    if (initialised_)
-        ::mysql_close(&mysql_);
-}
-
-
-void MySQLDbConnection::init(const std::string &database_name, const std::string &user, const std::string &passwd,
-                             const std::string &host, const unsigned port, const DbConnection::Charset charset,
-                             const DbConnection::TimeZone time_zone)
-{
-    initialised_ = false;
-
-    if (::mysql_init(&mysql_) == nullptr)
-        LOG_ERROR("mysql_init() failed! (1)");
-
-    if (::mysql_real_connect(&mysql_, host.c_str(), user.c_str(), passwd.c_str(), database_name.c_str(), port,
-                             /* unix_socket = */nullptr, /* client_flag = */CLIENT_MULTI_STATEMENTS) == nullptr)
-
-        // password is intentionally omitted here!
-        LOG_ERROR("mysql_real_connect() failed! (" + getLastErrorMessage()
-                  + ", host=\"" + host + "\", user=\"" + user + "\", passwd=\"" + passwd + "\", database_name=\""
-                  + database_name + "\", port=" + std::to_string(port) + ")");
-    if (::mysql_set_character_set(&mysql_, (charset == UTF8MB4) ? "utf8mb4" : "utf8") != 0)
-        LOG_ERROR("mysql_set_character_set() failed! (" + getLastErrorMessage() + ")");
-    errno = 0; // Don't ask unless you want to cry!
-
-    initialised_ = true;
-    setTimeZone(time_zone);
-    database_name_ = database_name;
-    user_ = user;
-    passwd_ = passwd;
-    host_ = host;
-    port_ = port;
-    charset_ = charset;
-}
-
-
-void MySQLDbConnection::init(const std::string &user, const std::string &passwd, const std::string &host, const unsigned port,
-                             const DbConnection::Charset charset, const DbConnection::TimeZone time_zone)
-{
-    initialised_ = false;
-
-    if (::mysql_init(&mysql_) == nullptr)
-        LOG_ERROR("mysql_init() failed! (2)");
-
-    if (::mysql_real_connect(&mysql_, host.c_str(), user.c_str(), passwd.c_str(), nullptr, port,
-                             /* unix_socket = */nullptr, /* client_flag = */CLIENT_MULTI_STATEMENTS) == nullptr)
-        LOG_ERROR("::mysql_real_connect() failed! (" + getLastErrorMessage() + ")");
-    if (::mysql_set_character_set(&mysql_, CharsetToString(charset).c_str()) != 0)
-        LOG_ERROR("::mysql_set_character_set() failed! (" + getLastErrorMessage() + ")");
-
-    initialised_ = true;
-    setTimeZone(time_zone);
-    user_ = user;
-    passwd_ = passwd;
-    host_ = host;
-    port_ = port;
-    charset_ = charset;
-}
-
-
-void MySQLDbConnection::setTimeZone(const TimeZone time_zone) {
-    switch (time_zone) {
-    case TZ_SYSTEM:
-        /* Default => Do nothing! */
-        break;
-    case TZ_UTC:
-        if (::mysql_query(&mysql_, "SET time_zone = '+0:00'") != 0)
-            LOG_ERROR("failed to set the connection time zone to UTC! (" + std::string(::mysql_error(&mysql_)) + ")");
-        break;
-    }
-    time_zone_ = time_zone;
-}
-
-
-bool MySQLDbConnection::query(const std::string &query_statement) {
-    if (MiscUtil::SafeGetEnv("UTIL_LOG_DEBUG") == "true")
-        FileUtil::AppendString(UBTools::GetTueFindLogPath() + "sql_debug.log",
-                               std::string(::program_invocation_name) + ": " +  query_statement + '\n');
-
-    const auto statements(SplitMySQLStatements(query_statement));
-    for (const auto &statement : statements) {
-        if (::mysql_query(&mysql_, statement.c_str()) != 0) {
-            LOG_WARNING("Could not successfully execute statement \"" + statement + "\": SQL error code:"
-                        + std::to_string(::mysql_errno(&mysql_)));
-            return false;
-        }
-    }
-    return true;
-}
-
-
-namespace {
-
-
-// Splits a compound Sqlite SQL statement into individual statements and eliminates comments.
-void SplitSqliteStatements(const std::string &compound_statement, std::vector<std::string> * const individual_statements) {
-    ParseState parse_state(NORMAL);
-    std::string current_statement;
-    char last_ch('\0');
-    for (const char ch : compound_statement) {
-        if (parse_state == IN_DOUBLE_DASH_COMMENT) {
-            if (ch == '\n')
-                parse_state = NORMAL;
-        } else if (parse_state == IN_C_STYLE_COMMENT) {
-            if (ch == '/' and last_ch == '*')
-                parse_state = NORMAL;
-        } else if (parse_state == IN_STRING_CONSTANT) {
-            if (ch == '\'')
-                parse_state = NORMAL;
-            current_statement += ch;
-        } else { // state == NORMAL
-            if (ch == '-' and last_ch == '-') {
-                current_statement.resize(current_statement.size() - 1); // Remove the trailing dash.
-                parse_state = IN_DOUBLE_DASH_COMMENT;
-            } else if (ch == '*' and last_ch == '/') {
-                current_statement.resize(current_statement.size() - 1); // Remove the trailing slash.
-                parse_state = IN_C_STYLE_COMMENT;
-            } else if (ch == ';') {
-                StringUtil::TrimWhite(&current_statement);
-                if (not current_statement.empty()) {
-                    current_statement += ';';
-                    AddStatement(current_statement, individual_statements);
-                    current_statement.clear();
-                }
-            } else if (ch == '\'') {
-                parse_state = IN_STRING_CONSTANT;
-                current_statement += ch;
-            } else
-                current_statement += ch;
-        }
-
-        last_ch = ch;
-    }
-
-    if (parse_state == IN_C_STYLE_COMMENT)
-        LOG_ERROR("unterminated C-style comment in SQL statement sequence: \"" + compound_statement + "\"!");
-    else if (parse_state == IN_STRING_CONSTANT)
-        LOG_ERROR("unterminated string constant in SQL statement sequence: \"" + compound_statement + "\"!");
-    else if (parse_state == NORMAL) {
-        StringUtil::TrimWhite(&current_statement);
-        if (not current_statement.empty()) {
-            current_statement += ';';
-            AddStatement(current_statement, individual_statements);
-        }
-    }
-}
-
-
-} // unnamed namespace
-
-
-bool MySQLDbConnection::queryFile(const std::string &filename) {
-    std::string statements;
-    if (not FileUtil::ReadString(filename, &statements))
-        return false;
-
-    const auto individual_statements(SplitMySQLStatements(statements));
-    for (const auto &statement : individual_statements) {
-        if (::mysql_query(&mysql_, statement.c_str()) != 0) {
-            LOG_WARNING("Could not successfully execute statement \"" + statement + "\": SQL error code:"
-                        + std::to_string(::mysql_errno(&mysql_)));
-            return false;
-        }
-    }
-
-    return true;
-}
-
-
-bool MySQLDbConnection::backup(const std::string &/*output_filename*/, std::string * const /*err_msg*/) {
-    LOG_ERROR("not yet implemented!");
-}
-
-
-DbResultSet MySQLDbConnection::getLastResultSet() {
-    MYSQL_RES * const result_set(::mysql_store_result(&mysql_));
-    if (result_set == nullptr)
-        LOG_ERROR("::mysql_store_result() failed! (" + getLastErrorMessage() + ")");
-
-    return DbResultSet(new MySQLResultSet(result_set));
-}
-
-
-std::string MySQLDbConnection::escapeString(const std::string &unescaped_string, const bool add_quotes,
-                                            const bool return_null_on_empty_string)
-{
-    if (unescaped_string.empty() and return_null_on_empty_string)
-        return "NULL";
-
-    char * const buffer(reinterpret_cast<char *>(std::malloc(unescaped_string.size() * 2 + 1)));
-    size_t escaped_length;
-
-    escaped_length = ::mysql_real_escape_string(&mysql_, buffer, unescaped_string.data(), unescaped_string.size());
-
-    std::string escaped_string;
-    escaped_string.reserve(escaped_length + (add_quotes ? 2 : 0));
-    if (add_quotes)
-        escaped_string += '\'';
-
-    escaped_string.append(buffer, escaped_length);
-    std::free(buffer);
-
-    if (add_quotes)
-        escaped_string += '\'';
-    return escaped_string;
-}
-
-
-bool MySQLDbConnection::tableExists(const std::string &database_name, const std::string &table_name) {
-    queryOrDie("SELECT EXISTS(SELECT * FROM information_schema.tables WHERE table_schema = '" + escapeString(database_name)
-               + "' AND table_name = '" + escapeString(table_name) + "')");
-    DbResultSet result_set(getLastResultSet());
-    return result_set.getNextRow()[0] == "1";
-}
-
-
-Sqlite3DbConnection::Sqlite3DbConnection(const std::string &database_path, const OpenMode open_mode)
-    : stmt_handle_(nullptr), database_path_(database_path)
-{
-    int flags(0);
-    switch (open_mode) {
-    case READONLY:
-        flags = SQLITE_OPEN_READONLY | SQLITE_OPEN_FULLMUTEX;
-        break;
-    case READWRITE:
-        flags = SQLITE_OPEN_READWRITE | SQLITE_OPEN_FULLMUTEX;
-        break;
-    case CREATE:
-        flags = SQLITE_OPEN_READWRITE | SQLITE_OPEN_CREATE | SQLITE_OPEN_FULLMUTEX;
-        break;
-    }
-
-    if (::sqlite3_open_v2(database_path.c_str(), &sqlite3_, flags, nullptr) != SQLITE_OK)
-        LOG_ERROR("failed to create or open an Sqlite3 database with path \"" + database_path + "\"!");
-    errno = 0; // It seems that sqlite3_open_v2 internally tries something that fails but doesn't clear errno afterwards.
-    initialised_ = true;
-}
-
-
-Sqlite3DbConnection::~Sqlite3DbConnection() {
-    if (initialised_) {
-        if (stmt_handle_ != nullptr) {
-            const int result_code(::sqlite3_finalize(stmt_handle_));
-            if (result_code != SQLITE_OK)
-                LOG_ERROR("failed to finalise an Sqlite3 statement! (" + getLastErrorMessage() + ", code was "
-                          + std::to_string(result_code) + ")");
-        }
-        if (::sqlite3_close(sqlite3_) != SQLITE_OK)
-            LOG_ERROR("failed to cleanly close an Sqlite3 database!");
-    }
-}
-
-
-bool Sqlite3DbConnection::query(const std::string &query_statement) {
-    if (MiscUtil::SafeGetEnv("UTIL_LOG_DEBUG") == "true")
-        FileUtil::AppendString(UBTools::GetTueFindLogPath() + "sql_debug.log",
-                               std::string(::program_invocation_name) + ": " +  query_statement + '\n');
-
-    if (stmt_handle_ != nullptr) {
-        const int result_code(::sqlite3_finalize(stmt_handle_));
-        if (result_code != SQLITE_OK) {
-            LOG_WARNING("failed to finalise an Sqlite3 statement! (" + getLastErrorMessage() + ", code was "
-                        + std::to_string(result_code) + ")");
-            return false;
-        }
-    }
-
-    const char *rest;
-    if (::sqlite3_prepare_v2(sqlite3_, query_statement.c_str(), query_statement.length(), &stmt_handle_, &rest) != SQLITE_OK)
-        return false;
-    if (rest != nullptr and *rest != '\0')
-        LOG_ERROR("junk after SQL statement (" + query_statement + "): \"" + std::string(rest) + "\"!");
-    switch (::sqlite3_step(stmt_handle_)) {
-    case SQLITE_DONE:
-    case SQLITE_OK:
-        if (::sqlite3_finalize(stmt_handle_) != SQLITE_OK)
-            LOG_ERROR("failed to finalise an Sqlite3 statement! (" + getLastErrorMessage() + ")");
-        stmt_handle_ = nullptr;
-        break;
-    case SQLITE_ROW:
-        break;
-    default:
-        return false;
-    }
-
-    return true;
-}
-
-
-bool Sqlite3DbConnection::queryFile(const std::string &filename) {
-    std::string statements;
-    if (not FileUtil::ReadString(filename, &statements))
-        return false;
-
-    std::vector<std::string> individual_statements;
-    SplitSqliteStatements(statements, &individual_statements);
-    for (const auto &statement : individual_statements) {
-        if (not query(statement))
-            return false;
-    }
-
-    return true;
-}
-
-
-bool Sqlite3DbConnection::backup(const std::string &output_filename, std::string * const err_msg) {
-    sqlite3 *sqlite3_backup_file;
-    int return_code;
-    if ((return_code = ::sqlite3_open(output_filename.c_str(), &sqlite3_backup_file)) != SQLITE_OK) {
-        *err_msg = "failed to create backup to \"" + output_filename + "\": "
-                   + std::string(::sqlite3_errmsg(sqlite3_backup_file));
-        return false;
-    }
-
-    sqlite3_backup *backup_handle(::sqlite3_backup_init(sqlite3_backup_file, "main", sqlite3_, "main"));
-    if (backup_handle == nullptr) {
-        ::sqlite3_close(sqlite3_backup_file);
-        *err_msg = "failed to initialize Sqlite3 backup to \"" + output_filename
-                   + "\": there is already a read or read-write transaction open on the destination database!";
-        return false;
-    }
-
-    const int DATABASE_PAGE_COUNT(500); // How many pages to copy to the backup at each iteration.
-    const int SLEEP_INTERVAL(100);      // in milliseconds.
-    bool backup_incomplete;
-    do {
-        return_code = ::sqlite3_backup_step(backup_handle, DATABASE_PAGE_COUNT);
-        backup_incomplete = return_code == SQLITE_OK or return_code == SQLITE_BUSY or return_code == SQLITE_LOCKED;
-        if (backup_incomplete)
-            ::sqlite3_sleep(SLEEP_INTERVAL);
-    } while (backup_incomplete);
-    ::sqlite3_backup_finish(backup_handle);
-
-    return_code = ::sqlite3_errcode(sqlite3_backup_file);
-    ::sqlite3_close(sqlite3_backup_file);
-    if (return_code != SQLITE_OK) {
-        *err_msg = "an error occurred during the backup to \"" + output_filename + "\": "
-                   + std::string(::sqlite3_errmsg(sqlite3_backup_file));
-        return false;
-    }
-
-    return true;
-}
-
-
-DbResultSet Sqlite3DbConnection::getLastResultSet() {
-    const auto temp_handle(stmt_handle_);
-    stmt_handle_ = nullptr;
-    return DbResultSet(new Sqlite3ResultSet(temp_handle));
-}
-
-
-std::string Sqlite3DbConnection::escapeString(const std::string &unescaped_string, const bool add_quotes,
-                                              const bool return_null_on_empty_string)
-{
-    if (unescaped_string.empty() and return_null_on_empty_string)
-        return "NULL";
-
-    char * const buffer(reinterpret_cast<char *>(std::malloc(unescaped_string.size() * 2 + 1)));
-    size_t escaped_length;
-
-    char *cp(buffer);
-    for (char ch : unescaped_string) {
-        if (ch == '\'')
-            *cp++ = '\'';
-        *cp++ = ch;
-    }
-
-    escaped_length = cp - buffer;
-
-    std::string escaped_string;
-    escaped_string.reserve(escaped_length + (add_quotes ? 2 : 0));
-    if (add_quotes)
-        escaped_string += '\'';
-
-    escaped_string.append(buffer, escaped_length);
-    std::free(buffer);
-
-    if (add_quotes)
-        escaped_string += '\'';
-    return escaped_string;
-}
-
-
-bool Sqlite3DbConnection::tableExists(const std::string &database_name, const std::string &table_name) {
-    Sqlite3DbConnection connection(database_name, READONLY);
-    connection.queryOrDie("SELECT name FROM sqlite_master WHERE type='table' AND name='"
-                          + connection.escapeString(table_name) + "'");
-    return not connection.getLastResultSet().empty();
-}
-
-
-PostgresDbConnection::~PostgresDbConnection() {
-    if (initialised_) {
-        ::PQfinish(pg_conn_);
-        pg_conn_ = nullptr;
-    }
-}
-
-
-unsigned PostgresDbConnection::getNoOfAffectedRows() const {
-    if (pg_result_ == nullptr)
-        LOG_ERROR("no result set available!");
-    const char * const no_of_affected_rows_as_string(::PQcmdTuples(pg_result_));
-    if (unlikely(*no_of_affected_rows_as_string == '\0'))
-        return 0;
-    return std::strtoul(no_of_affected_rows_as_string, nullptr, 10);
-}
-
-
-int PostgresDbConnection::getLastErrorCode() const {
-    // Postgres uses alphanumeric error codes!
-    LOG_ERROR("cannot be implemented for Postgres!");
-}
-
-
-bool PostgresDbConnection::query(const std::string &query_statement) {
-    if (MiscUtil::SafeGetEnv("UTIL_LOG_DEBUG") == "true")
-        FileUtil::AppendString(UBTools::GetTueFindLogPath() + "sql_debug.log",
-                               std::string(::program_invocation_name) + ": " +  query_statement + '\n');
-
-    pg_result_ = ::PQexec(pg_conn_, query_statement.c_str());
-    if (pg_result_ == nullptr)
-        LOG_ERROR("out of memory or failure to send the query to the server!");
-
-    switch (::PQresultStatus(pg_result_)) {
-    case PGRES_BAD_RESPONSE:
-    case PGRES_FATAL_ERROR:
-        ::PQclear(pg_result_);
-        pg_result_ = nullptr;
-        return false;
-    default:
-        return true;
-    }
-}
-
-
-bool PostgresDbConnection::queryFile(const std::string &filename) {
-    std::string statements;
-    if (not FileUtil::ReadString(filename, &statements))
-        return false;
-
-    const auto individual_statements(SplitPostgresStatements(statements));
-    for (const auto &statement : individual_statements) {
-        if (pg_result_ != nullptr)
-            ::PQclear(pg_result_);
-        pg_result_ = ::PQexec(pg_conn_, statement.c_str());
-        if (pg_result_ == nullptr)
-            LOG_ERROR("out of memory or failure to send the query to the server!");
-
-        switch (::PQresultStatus(pg_result_)) {
-        case PGRES_BAD_RESPONSE:
-        case PGRES_FATAL_ERROR:
-            ::PQclear(pg_result_);
-            pg_result_ = nullptr;
-            return false;
-        default:
-            /* do nothing */;
-        }
-    }
-
-    return true;
-}
-
-
-bool PostgresDbConnection::backup(const std::string &/*output_filename*/, std::string * const /*err_msg*/) {
-    LOG_ERROR("not yet implemented!");
-}
-
-
-DbResultSet PostgresDbConnection::getLastResultSet() {
-    const auto temp_pg_result = pg_result_;
-    pg_result_ = nullptr;
-    return DbResultSet(new PostgresResultSet(temp_pg_result));
-}
-
-
-std::string PostgresDbConnection::escapeString(const std::string &unescaped_string, const bool add_quotes,
-                                               const bool return_null_on_empty_string)
-{
-    if (unescaped_string.empty() and return_null_on_empty_string)
-        return "NULL";
-
-    char * const escaped_c_string(::PQescapeLiteral(pg_conn_, unescaped_string.c_str(), unescaped_string.size()));
-    if (unlikely(escaped_c_string == nullptr))
-        LOG_ERROR("failed to escape string! (" + getLastErrorMessage() + ")");
-
-    std::string escaped_string;
-    if (add_quotes)
-        escaped_string = "'" + std::string(escaped_c_string) + "'";
-    else
-        escaped_string = escaped_c_string;
-    ::PQfreemem(reinterpret_cast<void *>(escaped_c_string));
-
-    return escaped_string;
-}
-
-
-bool PostgresDbConnection::tableExists(const std::string &database_name, const std::string &table_name) {
-    queryOrDie("SELECT EXISTS (SELECT FROM pg_tables WHERE schemaname = '" + database_name
-               + "' AND tablename = '" + table_name + "'");
-    DbResultSet result_set(getLastResultSet());
-    return result_set.getNextRow()[0] == "t";
 }
